show unread notifications if some redisNotifications are not opened
